Temporarly pin branches for seed builds.
In order to create the initial omnibus-toolchain we need to use the
existing branches which contain all the stuff in omnibus and
omnibus-software which we intend to mostly remove. Once the initial
seed builds have been created, the required Windows toolchain branches
have been merged, and the slaves are updated to use the new toolchain,
we can remove the branch pinnings and use master.

Signed-off-by: Ryan Hass <rhass@users.noreply.github.com>