Force rebuild to get new gcc in msys2 environment
Version 1.1.66 fails to compile dep-selector-libgecode, a core
dependency of many Chef projects.  The failure is the result of a
macro conflict throughout the C/C++ standard library header files.

The problem can reproduced by attempting to compile the following
simple C++ program:

    #include <windows.h>
    #include <istream>

    int main(void) {
        return 0;
    }

This problem was introduced upstream, likely in the following commit:

https://sourceforge.net/p/mingw-w64/mingw-w64/ci/b7f44b677f43c3bb5f771ff5a430206de040c48c/

(This commit made it more likely to hit the problem, it was probably
possible to hit the issue before with different combinations of
includes.)

The package build scripts were modified to include a patch which
resolves the issue:

https://github.com/Alexpux/MINGW-packages/pull/2464

This fix appears in gcc-6.3.0-3 and related packages.  Version 1.1.66
shipped with gcc-6.3.0-2 and does not have this important fix.  A
rebuild should pull in the latest version and resolve the issue.

Signed-off-by: Steven Danna <steve@chef.io>