Add Finder auth flow helper script
This script will force MacOS into the authentication flow used by the
Omnibus DMG compressor. By executing this script, we can pre-authorize
machines so that the Omnibus builds do not hang and fail.

This resulting authentication flow needs to be accepted through the UI.
There is no way to programatically accept the auth flow at this time.

Signed-off-by: Tom Duffield <tom@chef.io>