Add berkshelf to omnibus toolchain
This is used by Server, Supermarket, Manage, and Backend for the build
time installation of gems. If we move it to toolchain then we can avoid
installing a ton of deps into our packages and also unblock supermarket
and manage which are totally broken trying to install Berkshelf right
now.

Signed-off-by: Tim Smith <tsmith@chef.io>